pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-poison-defense"
version = "0.1.0"
description = "Reduced-scale end-to-end implementation of a PoisonedRAG defense study"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.40.0",
  "PyYAML>=6.0.1",
  "numpy>=1.26.0",
  "scikit-learn>=1.4.0",
  "pandas>=2.2.0",
  "sentencepiece>=0.2.0",
  "transformers>=4.44.0",
  "faiss-cpu>=1.8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
